+++ CHANGELOG.md
@@ -9,11 +9,13 @@ Breaking Changes:
 * 移除utils.config.fetchCoinCfg
 * 移除utils.config.fetchAllCoinNames
 * 重构setuper，仅将default钱包中的数据缓存到内存config
+* 移除crypto.fetchPubKey，仅使用新方法fetchPublicKeys
 
 Features:
 
 * 新增Wallet模型，支持多热钱包模式
 * 为Application配置Wallet参数
+* crypto的internal签名支持以timestamp + secret的形式做为私钥
 
 Improvement: